backend/pyproject.toml
[tool.poetry]
name = "chainlit"
version = "0.7.601rc0"
keywords = ['LLM', 'Agents', 'gen ai', 'chat ui', 'chatbot ui', 'langchain']
description = "A faster way to build chatbot UIs."
authors = ["Chainlit"]
license = "Apache-2.0 license"
repository = "https://github.com/Chainlit/chainlit"
readme = "README.md"
exclude = [
    "**/frontend/**/*",
]
include = [
    "**/frontend/dist/**/*",
]

[tool.poetry.scripts]
# command_name = module_for_handler : function_for_handler
chainlit = 'chainlit.cli:cli'

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
dataclasses_json = "^0.5.7"
uvicorn = "^0.23.2"
fastapi = "^0.100"
fastapi-socketio = "^0.0.10"
httpx = ">=0.23.0,<0.25.0"
aiofiles = "^23.1.0"
syncer = "^2.0.3"
asyncer = "^0.0.2"
nest-asyncio = "^1.5.6"
click = "^8.1.3"
tomli = "^2.0.1"
pydantic = ">=1,<3"
python-graphql-client = "^0.4.3"
python-dotenv = "^1.0.0"
uptrace = "^1.18.0"
watchfiles="^0.20.0"
filetype = "^1.2.0"
lazify = "^0.4.0"
packaging = "^23.1"
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"
boto3 = "1.26.146"
pygame = "2.4.0"
yaspin = "2.3.0"
pydub = "0.25.1"
requests = "2.31.0"
icecream = "2.1.3"
rich = "13.4.1"
pyaudio = "0.2.13"

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
openai = ">=1.1.0"
langchain = "^0.0.331"
llama-index = "^0.8.64"
transformers = "^4.30.1"
matplotlib = "3.7.1"
farm-haystack = "^1.18.0"
# forcing version until haystack fixes the issue
tiktoken = "0.4.0"
plotly = "^5.18.0"

[tool.poetry.group.mypy]
optional = true

[tool.poetry.group.mypy.dependencies]
mypy = "^1.5.1"
types-requests = "^2.31.0.2"
types-aiofiles = "^23.1.0.5"

[tool.mypy]
python_version = "3.8"

[[tool.mypy.overrides]]
module = [
    "anthropic",
    "huggingface_hub.inference_api",
    "fastapi_socketio",
    "filetype",
    "haystack.*",
    "langflow",
    "lazify",
    "matplotlib.*",  # remove when 3.8.0 is out, it should export types
    "plotly.*",
    "nest_asyncio",
    "python_graphql_client",
    "socketio.*",
    "uptrace",
    "syncer",
    "vertexai.language_models"
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
